Add generic locksutil shards + helpers
This ended up not being used for external keys, but it is a nice
improvement nevertheless.
